Create forestTree
Assignment submission for Week6

Generate a random list of integers.  
Show the binary heap tree resulting from inserting the integers on the list one at a time.
Using the list from the previous question, show the binary heap tree resulting from the list as a parameter to the buildHeap method.  
Show both the tree and list form.